requirements.txt
uvicorn
seaborn
numpy
dash
plotly.express
pandas